we will code specificity after finnishing base class
